+++ .babelrc
@@ -1,3 +1,10 @@
 {
-  "presets": ["@babel/preset-react"]
+  "presets": [
+    "@babel/preset-react",
+    ["@babel/preset-env", {
+      "targets": {
+        "node": "current"
+      }
+    }]
+  ]
 }
