+++ .eslintrc.yml
@@ -2,6 +2,7 @@ env:
   es6: true
   browser: true
   commonjs: true
+  jest: true
 
 extends: ['htmlacademy/es6', 'plugin:react/recommended']
