+++ package.json
@@ -5,18 +5,22 @@
   "description": "Личный проект «Шесть городов» от HTML Academy",
   "scripts": {
     "eslint": "eslint --ext .jsx --ext .js src/",
-    "test": "npm run eslint",
+    "test": "npm run eslint && npm run test.jest",
     "build": "webpack --mode production",
-    "start": "webpack-dev-server"
+    "start": "webpack-dev-server",
+    "test.jest": "jest"
   },
   "devDependencies": {
     "@babel/core": "7.10.2",
     "@babel/preset-env": "7.10.2",
     "@babel/preset-react": "7.10.1",
+    "babel-jest": "26.0.1",
     "babel-loader": "8.1.0",
     "eslint": "5.13.0",
     "eslint-config-htmlacademy": "0.5.2",
     "eslint-plugin-react": "7.20.0",
+    "jest": "26.0.1",
+    "react-test-renderer": "16.13.1",
     "shortid": "2.2.15",
     "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
